Kotlin conversion: Manually fix Kotlin conversion errors (#88). Add kotlin-reflect for reflection support. Add maven-surefire-plugin to support running JUnit 5 tests.

+++ renderer/src/main/kotlin/io/sparkled/renderer/easing/EasingFunction.kt
@@ -6,13 +6,11 @@ interface EasingFunction {
 
     /**
      * @param easing         The easing configuration, including parameters
-     * *
      * @param currentFrame   The current animation effect frame being rendered
-     * *
      * @param frameCount     The number of frames in the animation effect (implementing functions will likely subtract 1
-     * *                       from this value, as currentFrame is zero-based)
-     * *
+     *                       from this value, as currentFrame is zero-based)
+     *
      * @return A value between 0 and 1 describing the overall progress of the animation effect
      */
-    fun getProgress(easing: Easing, currentFrame: Float, frameCount: Float): Float
+    fun getProgress(easing: Easing, currentFrame: Int, frameCount: Int): Float
 }
